Migration DuckDB, étape 2 : le pipeline ne parle plus à Postgres
Zéro `psycopg`, zéro `pg_*`, zéro `Jsonb` dans le pipeline. 26 modules
convertis, 13 chargements en masse réécrits, `db.py`, `sauvegarde.py` et
`retablissement.py` supprimés — la sauvegarde est le fichier versionné par
le bucket, et il n'y a plus de base à remettre en écriture.

Le portage était mécanique une fois vérifié qu'aucun `%s` de ce dépôt n'est
hors SQL, et que `with conn.cursor()` se comporte pareil sous DuckDB. Ce
qui ne l'était pas :

- `maintenance.py` lançait un `vacuum full` table par table — l'opération
  qui, sur une base saturée, réclame justement l'espace qu'elle doit
  libérer. Elle devient un checkpoint, en une instruction et sans verrou.
- `limites.py` gardait un plafond de 470 Mo, contrainte d'un plan gratuit.
  L'entrepôt est un fichier dans un bucket : le garde-fou reste, à 8 Go,
  mais comme seuil d'invraisemblance — un connecteur qui produirait cent
  fois le volume attendu est en défaut. Il n'est plus arbitre budgétaire,
  rôle qu'il n'aurait jamais dû avoir.
- `jsonb_agg`/`jsonb_build_object` deviennent `json_group_array`/
  `json_object`.

**Ce que la migration a fait apparaître, et qui est le vrai gain.** Les
dix-huit tests qui exigeaient un Postgres joignable étaient sautés partout
sauf dans un job dédié : ils ne vérifiaient donc à peu près jamais rien.
Ils tournent maintenant sur un fichier temporaire, en local et en CI, en
quatre secondes. Deux défauts sont sortis dès la première exécution :

1. La table de passage entre millésimes rendait **0,8 et 0,3 au lieu de
   0,75 et 0,25**. Le littéral `1.0` est un DECIMAL(2,1) pour DuckDB, et la
   multiplication en gardait l'échelle : une scission de commune
   réattribuait sa population de travers, sans erreur, sur toute
   comparaison inter-millésimes. Les parts sont désormais en `double`.
   C'est exactement le genre de corruption silencieuse que ce projet existe
   pour ne pas produire, et elle est passée à un cheveu de ne pas être vue.
2. DuckDB refuse un `rollback()` hors transaction là où psycopg
   l'acceptait. Ces appels disent tous « repars propre », jamais « il y
   avait forcément une transaction » : `entrepot.annuler()` les rend
   tolérants, et le dit.

**État : 175 tests passent, 8 échouent.** Ils échouent parce qu'ils
s'exécutent enfin, pas parce que le portage les a cassés — aucun ne
tournait avant. Ils sont listés dans la tâche de suivi et c'est la
prochaine étape :

  test_lineage_db : test_failed_run_is_traced
  test_publication : test_les_series_pays_gardent_leur_profondeur_entiere
  test_registry_db : test_sync_est_idempotent_et_couvre_les_jeux_p0
  test_revisions_db : deux tests d'archivage des révisions
  test_secu, test_securite : test_declarer_passe_les_contraintes_de_la_base

Rien du site n'est touché : aucun fichier de `site/` dans ce commit.

+++ infra/entrepot/schema.sql
@@ -373,9 +373,13 @@ create table if not exists fin.public_employment (
 -- table DuckDB — la récursion, elle, est identique.
 create or replace macro geo.passage(p_level, p_code, p_vintage, p_target) as table
 with recursive walk(code, vintage, share) as (
-    select p_code, p_vintage, 1.0
+    -- Les parts sont en double, pas en décimal. Le littéral `1.0` est un
+    -- DECIMAL(2,1) pour DuckDB : la multiplication en gardait l'échelle, et une
+    -- scission à 0,75 / 0,25 ressortait à 0,8 / 0,3 — une population réattribuée
+    -- de travers, sans erreur, sur toute comparaison inter-millésimes.
+    select p_code, p_vintage, 1.0::double
     union all
-    select h.to_code, h.to_vintage, w.share * coalesce(h.population_share, 1)
+    select h.to_code, h.to_vintage, w.share * coalesce(h.population_share, 1)::double
     from walk w
     join geo.geography_history h
       on h.from_level = p_level
